+++ src/core/mixins/learnerCreationMixin.js
@@ -1,6 +1,5 @@
 import get from 'lodash/get';
-import { clear, removeEmptyProps, formatPhoneForPayload, formatIdentity, removeDiacritics } from '@helpers/utils';
-import { dateDiff } from '@helpers/date';
+import { clear, removeEmptyProps, formatPhoneForPayload } from '@helpers/utils';
 import Users from '@api/Users';
 import Email from '@api/Email';
 import { TRAINEE, INTRA } from '@data/constants';
@@ -11,7 +10,7 @@ import { required, email, requiredIf } from 'vuelidate/lib/validators';
 export const learnerCreationMixin = {
   data () {
     const isClientInterface = !/\/ad\//.test(this.$router.currentRoute.path);
-    const isBlendedCourseProfile = !/\/learners\//.test(this.$router.currentRoute.path);
+    const isBlendedCourseProfile = !/\/learners/.test(this.$router.currentRoute.path);
 
     return {
       isClientInterface,
@@ -49,12 +48,21 @@ export const learnerCreationMixin = {
       this.firstStep = true;
       this.newLearner = { ...clear(this.newLearner) };
       this.$v.newLearner.$reset();
+      this.userAlreadyHasCompany = false;
     },
     formatUserPayload () {
       const payload = removeEmptyProps(this.newLearner);
       if (get(payload, 'contact.phone')) payload.contact.phone = formatPhoneForPayload(this.newLearner.contact.phone);
 
       return this.isClientInterface ? { ...payload, company: this.company._id } : payload;
+    },
+    setNewLearner (user) {
+      this.newLearner._id = user._id;
+      this.newLearner.identity = {
+        firstname: get(user, 'identity.firstname'),
+        lastname: get(user, 'identity.lastname'),
+      };
+      this.newLearner.contact = { phone: get(user, 'contact.phone') };
     },
     async nextStepLearnerCreationModal () {
       try {
@@ -71,29 +79,21 @@ export const learnerCreationMixin = {
 
           return this.goToCreationStep();
         }
+
         if (this.isBlendedCourseProfile) {
           const user = await Users.getById(userInfo.user._id);
-          this.newLearner = {
-            _id: user._id,
-            identity: { firstname: get(user, 'identity.firstname'), lastname: get(user, 'identity.lastname') },
-            contact: { phone: get(user, 'contact.phone') },
-            local: { email: get(user, 'local.email') },
-            company: get(user, 'company._id'),
-          };
-          this.userAlreadyHasCompany = !!get(user, 'company._id');
-          if (this.course.type === INTRA && !this.newLearner.company) this.newLearner.company = this.course.company._id;
-          if (this.course.type === INTRA && this.newLearner.company !== this.course.company._id) {
-            this.newLearner = { ...clear(this.newLearner) };
-            this.$v.newLearner.$reset();
-
-            return NotifyNegative('L\'apprenant(e) existe déjà et n\'est pas relié(e) à la bonne structure.');
-          }
-          if (get(user, 'role.client.name') === 'helper' && this.course.type === INTRA) {
-            this.newLearner = { ...clear(this.newLearner) };
-            this.$v.newLearner.$reset();
-
+          if (['helper', 'auxiliary_without_company'].includes(get(user, 'role.client.name')) &&
+            this.course.type === INTRA) {
             return NotifyNegative('Cette personne ne peut pas être ajoutée à la formation.');
           }
+          if (get(user, 'company._id')) this.newLearner.company = get(user, 'company._id');
+          if (this.course.type === INTRA && !this.newLearner.company) this.newLearner.company = this.course.company._id;
+          if (this.course.type === INTRA && this.newLearner.company !== this.course.company._id) {
+            return NotifyNegative('L\'apprenant(e) existe déjà et n\'est pas relié(e) à la bonne structure.');
+          }
+
+          this.setNewLearner(user);
+          this.userAlreadyHasCompany = !!get(user, 'company._id');
 
           return this.goToCreationStep();
         }
@@ -119,7 +119,6 @@ export const learnerCreationMixin = {
     },
     async createLearner () {
       try {
-        this.learnerCreationModalLoading = true;
         this.$v.newLearner.$touch();
         if (this.$v.newLearner.$error) return NotifyWarning('Champ(s) invalide(s).');
 
@@ -128,15 +127,10 @@ export const learnerCreationMixin = {
         NotifyPositive('Apprenant(e) ajouté(e) avec succès.');
 
         await this.sendWelcome();
-
-        this.learnerCreationModal = false;
-        await this.getLearnerList(this.isClientInterface ? this.company._id : null);
       } catch (e) {
         console.error(e);
         if (e.status === 409) return NotifyNegative(e.data.message);
         NotifyNegative('Erreur lors de l\'ajout de l\' apprenant(e).');
-      } finally {
-        this.learnerCreationModalLoading = false;
       }
     },
     async sendWelcome () {
@@ -146,41 +140,6 @@ export const learnerCreationMixin = {
       } catch (e) {
         console.error(e);
         NotifyNegative('Erreur lors de l\'envoi du mail.');
-      }
-    },
-    getDaysSinceLastActivityHistory (lastActivityHistory) {
-      if (!lastActivityHistory) return null;
-
-      return dateDiff(Date.now(), lastActivityHistory.updatedAt);
-    },
-    formatRow (user) {
-      const formattedName = formatIdentity(user.identity, 'FL');
-
-      return {
-        learner: {
-          _id: user._id,
-          fullName: formattedName,
-          lastname: user.identity.lastname,
-          picture: user.picture ? user.picture.link : null,
-          noDiacriticsName: removeDiacritics(formattedName),
-        },
-        company: user.company ? user.company.name : 'N/A',
-        blendedCoursesCount: user.blendedCoursesCount,
-        eLearningCoursesCount: user.eLearningCoursesCount,
-        activityHistoryCount: user.activityHistoryCount,
-        daysSinceLastActivityHistory: this.getDaysSinceLastActivityHistory(user.lastActivityHistory),
-      };
-    },
-    async getLearnerList (companyId = null) {
-      try {
-        this.tableLoading = true;
-        const learners = await Users.learnerList(companyId ? { company: companyId } : {});
-        this.learnerList = Object.freeze(learners.map(this.formatRow));
-      } catch (e) {
-        console.error(e);
-        this.learnerList = [];
-      } finally {
-        this.tableLoading = false;
       }
     },
   },
